refactor(types): restructure Category types and update imports
