+++ src/services/organization.ts
@@ -1,22 +1,15 @@
-import Category from '@/types/category-type';
+import api from './api';
+
+import { Category } from '@/types';
 import Organization from '@/types/organization-type';
 import Url from '@/types/url-type';
 
-import api from './api';
-
-async function getCategoryByOrganizationId(
-  organizationId: Organization['_id'],
-  page: number = 1,
-  limit: number = 10,
-) {
+async function getCategoryByOrganizationId(organizationId: Organization['_id'], page: number = 1, limit: number = 10) {
   try {
-    return (
-      await api.get(
-        `v1/api/organizations/${organizationId}/categories?page=${page}&limit=${limit}`,
-      )
-    ).data.payload;
+    return (await api.get(`v1/api/organizations/${organizationId}/categories?page=${page}&limit=${limit}`)).data
+      .payload;
   } catch (e: any) {
-    console.log(e.message);
+    console.error(e.message);
   }
 }
 
@@ -28,12 +21,10 @@ async function searchCategoryByOrganizationId(
 ) {
   try {
     return (
-      await api.get(
-        `v1/api/organizations/${organizationId}/categories/search?page=${page}&limit=${limit}&q=${query}`,
-      )
+      await api.get(`v1/api/organizations/${organizationId}/categories/search?page=${page}&limit=${limit}&q=${query}`)
     ).data.payload;
   } catch (e: any) {
-    console.log(e.message);
+    console.error(e.message);
   }
 }
 
@@ -42,14 +33,9 @@ async function addUrlToCategories(
   payload: { url: Url['_id']; categories: Category['_id'][] },
 ) {
   try {
-    return (
-      await api.patch(
-        `/v1/api/organizations/${organizationId}/categories/add-url`,
-        payload,
-      )
-    ).data.payload;
+    return (await api.patch(`/v1/api/organizations/${organizationId}/categories/add-url`, payload)).data.payload;
   } catch (e: any) {
-    console.log(e.message);
+    console.error(e.message);
   }
 }
