+++ src/types/Category.ts
@@ -1,3 +1,14 @@
+import Organization from '@/types/organization-type';
+import Url from '@/types/url-type';
+
+export type Category = {
+  _id: string;
+  name: string;
+  color: string;
+  organization: Organization['_id'];
+  urls: Url['_id'][];
+};
+
 export enum CategoryColor {
   RED = '#e30425',
   BLUE = '#84d2f4',
@@ -5,5 +16,3 @@ export enum CategoryColor {
   GREEN = '#48b24f',
   PURPLE = '#aaabcd',
 }
-
-// TODO: Merge category-type.ts in next PR
