+++ src/types/shorten-input-field-type.ts
@@ -1,5 +1,5 @@
-import Category from './category-type';
-import Organization from './organization-type';
+import { Category } from '@/types';
+import Organization from '@/types/organization-type';
 
 type ShortenInputFieldType = string | Organization | Category;
